+++ engines/mads/madsv2/phantom/rooms/room105.h
@@ -37,10 +37,10 @@ namespace Rooms {
 
 typedef struct {        /* Room local variables */
 
-	int sprite[15];       /* Sprite series handles */
-	int sequence[15];     /* Sequence handles      */
-	int animation[4];     /* Animation handles     */
-	int column_prop;      /* Dynamic HS for column_prop */
+	int16 sprite[15];       /* Sprite series handles */
+	int16 sequence[15];     /* Sequence handles      */
+	int16 animation[4];     /* Animation handles     */
+	int16 column_prop;      /* Dynamic HS for column_prop */
 
 } Scratch;
